Workaround links in subdirs
If a link starts with ../, ignore that part.

If a link is not found in RELATIONS, find a key that ends with
slash + link name.

This **will** break once we have files with identical filenames in
different directories, but it works now.

Proper fix will need to pass current file path to the
fix_relative_links() function and also to modify RelLinkerFakeDict.

Let's not go there just yet.